+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Standalone script 'PrivateMethodAccess.js'
+- Variant script 'AddUrlParams.js'
 ### Changed
 - Add cautionary note to help and readme.
 ### Fixed